Add compilation settings
There are two sets of settings, "file", hooked up to eval.one, and
"project", hooked up to eval.  This required a significant rewriring of
the eval process; notably, saving no longer triggers an eval by itself
(since we don't know which to use).  Settings are stored in
~/.config/litexrc and in .litexrc in each directory.

One of these settings is the PDF file produced, so it's no longer so
obvious which to use for forward syncing.  After compilation, we store
the file name in the editor and use that.  Before any compilations, we
use the filename from either the "file" or the "project" settings, if it
exists; otherwise, we bail out.

Ref #3.