Remove "directory" from settings
It was untested, and I don't think it would have worked in a
particularly useful way.  Instead, we take the directory of the filename
to be the working directory.

Also, simplify the figuring of the pdfname in get-config-from-settings,
since we should always have a settings value for "outputname" now that
synctex is handled elsewhere.