Render PDF page by page
There's a relatively small penalty for this, and it allows us to display
pages as soon as they're ready.  Right now, we just do them in order,
but we could modify this to start with the displayed page, or to do
just-in-time rendering.

Note that we are now setting the image size directly, rather than using
CSS zoom.  A high-res render with the correct image size looks a bit
clearer than the same image with a small size and compensating zoom.

Ref #4.